requirements.txt
google-genai>=0.3.0
python-dotenv>=1.0.1
pandas>=2.2.2
numpy>=2.0.0
tqdm>=4.66.4
pillow>=10.3.0
requests>=2.32.3
scikit-learn>=1.5.1
ipykernel>=6.29.5
jupyter>=1.0.0